+++ test/numo_test.rb
@@ -30,4 +30,10 @@ class NumoTest < Minitest::Test
     assert_kind_of Numo::RObject, s.to_numo
     assert_equal s.to_a, s.to_numo.to_a
   end
+
+  def test_data_frame
+    df = Polars::DataFrame.new({"a" => [1, 2, 3], "b" => ["one", "two", "three"]})
+    assert_kind_of Numo::RObject, df.to_numo
+    assert_equal [[1, "one"], [2, "two"], [3, "three"]], df.to_numo.to_a
+  end
 end